requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
phonenumbers==8.12.54
pytz==2022.2.1
sqlparse==0.4.2
